Implement geany-before-quit handler
During Geany shutdown, lots of things happen and since the main
loop is being executed manually to still process jsonrpc messages, it
can also process Geany events in a non-standard manner.

The signal helps to run the main loop in the place place in which not
many things in Geany happen yet so there's no such danger of processing
unwanted events from Geany.