Rework server launching to use Geany's spawn module instead of gsubprocess
Potentially this will allow the plugin to work under Windows.

There's a copy of spawn.c/h because lsp_spawn_async_with_pipes() is
needed and it's not part of Geany API (yet?).

Similarly, there's a copy of gunixinputstream and gunixoutputstream
that works on file descriptors which however isn't available
under Windows version of glib.